Moved base and modules directory under src (similar to INET).
This works better with makefile generation. The goal is to have a single
MiXiM dynamic library and one executable for the tests and the examples.
(Taken from Levente MiXiM-INET integration)

* Moved base and modules directory under src (similar to INET).

This works better with makefile generation. The goal is to have a single
MiXiM dynamic library and one executable for the tests and the examples.
(Taken from Levente MiXiM-INET integration)

* Moved base and modules directory under src (similar to INET).

This works better with makefile generation. The goal is to have a single
MiXiM dynamic library and one executable for the tests and the examples.
(Taken from Levente MiXiM-INET integration)

* Moved base and modules directory under src (similar to INET).

This works better with makefile generation. The goal is to have a single
MiXiM dynamic library and one executable for the tests and the examples.
(Taken from Levente MiXiM-INET integration)

Moved base and modules directory under src (similar to INET).

This works better with makefile generation. The goal is to have a single
MiXiM dynamic library and one executable for the tests and the examples.
(Taken from Levente MiXiM-INET integration)