Rework float comparison
When comparing large floats with FWMath::close(), the used epsilon is
not sufficient. The updated version scales the epsilon provided by the
standard library with the inputs to avoid this.
Additionally, the new function is using a namespace instead of a class
for scoping.

Veins::FWMath has been deprecated to indicate future discontinuation of
the API.