Fix ObstacleControl::finish overhead.
ObstacleControl::finish took a lot of time to clean up on finish (up to
minutes for large scenarios with many obstacles).
This was due to three nested loops in finish hhat called erase, which in
turn ran the same three nested loops to delete a single obstacle.

This patch simply adds an owning container for each covered obstacle.
It contains a unique_ptr as an ownder for each obstacle.
On finish, the container gets cleaned up, which in turn deletes the
Obstacle instances.

ObstacleControl::erase became a bit more complicated because it has to
find the owning unique_ptr for the Obstacle to delete.
However, ObstacleControl::erase has no longer a caller in the code and
was only left in because it is a public method.
(The only previous caller was ObstacleControl::finish).