Use automatic version number
To prevent an incorrect version number due to a missing update of the
value in the doxy.cfg, a variable for the version number is automatically filled.

A script is used to determine the current project version.
If the current version does not exactly match a tag, an abbreviation of
the current commit hash is used as version.

If the variable is not filled, e.g., the documentation is not generated
using the makefile target, the variable will be empty.