Person subscription manager added
This copies stuff from TraCIScenarioManager to PersonSubscriptionManager.

The person subscription manager expects creates an id list subscription
(in initalize()) and then manages which persons entered the simulation and
which left it. When a person enters the simulation it will add
a subscription for that specific person to subscribe to its position,
current street or its type.

Via the API of this class it is possible to request updates for persons
transported via the data exchange format TraCIPerson (added in previous
commits).