Traffic light subscription manager added
This copies stuff from TraCIScenarioManager to TrafficLightSubscriptionManager.

This is different from other subscription manager. It does not subscribe
to all traffic lights available. It instead subscribes to specific ones
that are added via the class api.

The API offers a way to request traffic light updates. The data type
was defined in previous commits.