Added ExecutiveSubscriptionManager
This is the "mastermind" who manages all subscription managers. It provides
them with newly arrived subscription data and ensures that everyone receives
subscription updates. If they do not receive an id list subscription the
manager will explicitly request an id list (can be switched off).

This class will be the central interface that other classes, that depend
on subscription, will work with. Most of the time the subscription manager
will simply delegate the calls to its managers.