Remove orientation attribute from Move object.
The orientation was a legacy attribute of the Move object.
On every update of its direction (of travel) attribute, it was overridden.
Only for the initial setting the two could differ.
No part of the Veins codebase uses the orientation now.
Instead, the direction of the Move object or the Heading should be used.
Thus, this commit removes the orientation altogether.