Change host identifier for VehicleObstacle.
Change from pointers to ChannelAccess to AntennaPosition.
This avoids potential crashes if the ChannelAccess (NIC) gets removed
while the message is still in flight.

An alternative would have been to just save the NIC component ids.
But that would have required adding a getter to the AntennaPosition.

This commit removes a trace message indicating the position of the
antennas of an obstacle in VehicleObstacleControl.