veins_inet: Send signals for INET visualization
To allow visualizing scenarios with dynamic modules,
INET needs to be notified of model changes.
This commit makes the VeinsInetManager emit these
required events.

Note that the signals used are already part of OMNeT++
and it is planned to move the notification objects there
as well.
Once this move is done as well and OMNeT features a full
model change notification mechanism, this is no longer
INET-specific should propably be implemented by the
TraCIScenarioManager instead.